% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serpstatr.R
\docType{package}
\name{serpstatr}
\alias{serpstatr-package}
\alias{serpstatr}
\title{serpstatr:  Serpstat API wrapper for R.}
\description{
The serpstatr package covers main Serpstat API methods. 
All the names of the functions start with \code{sst_} to reduce
the name conflict with other packages.
}
\details{
All the required arguments are checked by the API endpoint. So you would not
get an error in R but the response would contain \code{error} object.
}
\section{Functions naming}{

The names of the package functions start with \code{sst_sa_}
\itemize{
  \item \code{\link{sst_sa_database_info}}
  \item \code{\link{sst_sa_stats}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}
}

}
\author{
\strong{Maintainer}: Alex Danilin \email{alexnikdanilin@gmail.com}

Authors:
\itemize{
  \item Alex Danilin \email{alexnikdanilin@gmail.com}
}

}
\keyword{internal}
