% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_export}
\alias{sst_au_export}
\title{export}
\usage{
sst_au_export(
  report_id,
  export_type,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{export_type}{(required) Type of export file:

Type | Description
---------|----------
 mgxlsx | XLSX file
 mgxlsx_mfiles | Google Sheet 
 puppeter_pdf  | PDF file}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
export
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/p4x2q7dwptyvz-export}{here}.
}

