% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_audit}
\alias{sst_au_get_audit}
\title{getAudit}
\usage{
sst_au_get_audit(
  report_id,
  compare_report_id = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{compare_report_id}{(optional) `id` of your another report from your project to compare.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getAudit
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/xlw2titchn5w2-get-audit}{here}.
}

