% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_categories_statistic}
\alias{sst_au_get_categories_statistic}
\title{getCategoriesStatistic}
\usage{
sst_au_get_categories_statistic(
  report_id,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getCategoriesStatistic
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/0bpj0tlgitnfu-get-categories-statistic}{here}.
}

