% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_error_elements}
\alias{sst_au_get_error_elements}
\title{getErrorElements}
\usage{
sst_au_get_error_elements(
  report_id,
  compare_report_id,
  project_id,
  error_name,
  mode = NULL,
  limit = NULL,
  offset = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{compare_report_id}{(required) Another unique identifier for an audit report from the same project to compare}

\item{project_id}{(required)}

\item{error_name}{(required)}

\item{mode}{(optional) Error display mode:

Mode | Description
---------|----------
 all | all errors
 new | new errors
solved | fixed errors}

\item{limit}{(optional) count of returned items in response}

\item{offset}{(optional) batch number required for pagination}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getErrorElements
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/b9nl4a4f1vm7z-get-error-elements}{here}.
}

