% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_error_rows}
\alias{sst_au_get_error_rows}
\title{getErrorRows}
\usage{
sst_au_get_error_rows(
  report_id,
  error,
  compare_report_id = NULL,
  mode = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{error}{(required)}

\item{compare_report_id}{(optional) `id` of your another report from your project to compare.}

\item{mode}{(optional)  type of errors (all | new | solved) by default all}

\item{page}{(optional)}

\item{size}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getErrorRows
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/uuspflakzd2yi-get-error-rows}{here}.
}

