% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_history_by_count_error}
\alias{sst_au_get_history_by_count_error}
\title{getHistoryByCountError}
\usage{
sst_au_get_history_by_count_error(
  project_id,
  error_name,
  limit,
  offset,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{error_name}{(required)}

\item{limit}{(required) count of returned items in response}

\item{offset}{(required) batch number required for pagination}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getHistoryByCountError
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/uvufljcemtfnu-get-history-by-count-error}{here}.
}

