% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_pages_list}
\alias{sst_au_get_pages_list}
\title{getPagesList}
\usage{
sst_au_get_pages_list(
  limit = NULL,
  offset = NULL,
  team_member_id = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{limit}{(optional) count of returned items in response}

\item{offset}{(optional)  batch number required for pagination}

\item{team_member_id}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getPagesList
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/5xn8vauu8w6dk-get-pages-list}{here}.
}

