% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_report_without_details}
\alias{sst_au_get_report_without_details}
\title{getReportWithoutDetails}
\usage{
sst_au_get_report_without_details(
  report_id,
  compare_report_id = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{compare_report_id}{(optional) Another unique identifier for an audit report from the same project to compare}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getReportWithoutDetails
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/poj5fmggexd0y-get-report-without-details}{here}.
}

