% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_reports_list_by_page}
\alias{sst_au_get_reports_list_by_page}
\title{getReportsListByPage}
\usage{
sst_au_get_reports_list_by_page(
  page_id,
  limit = NULL,
  offset = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{page_id}{(required)}

\item{limit}{(optional) count of returned items in response}

\item{offset}{(optional) batch number required for pagination}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getReportsListByPage
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/dx0k6lh7g0z7c-get-reports-list-by-page}{here}.
}

