% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_get_sub_elements_by_crc}
\alias{sst_au_get_sub_elements_by_crc}
\title{getSubElementsByCrc}
\usage{
sst_au_get_sub_elements_by_crc(
  report_id,
  project_id,
  error_name,
  crc,
  compare_report_id = NULL,
  mode = NULL,
  limit = NULL,
  offset = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(required)}

\item{project_id}{(required)}

\item{error_name}{(required)}

\item{crc}{(required) URL crc}

\item{compare_report_id}{(optional)}

\item{mode}{(optional) Error display mode:

Mode | Description
---------|----------
 all | all errors
 new | new errors
solved | fixed errors}

\item{limit}{(optional) count of returned items in response}

\item{offset}{(optional) batch number required for pagination}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getSubElementsByCrc
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/arshynmdh3s1q-get-sub-elements-by-crc}{here}.
}

