% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_page_names}
\alias{sst_au_page_names}
\title{pageNames}
\usage{
sst_au_page_names(
  team_member_id = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{team_member_id}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
pageNames
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/mf6khxhf3h4dm-page-names}{here}.
}

