% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_scan}
\alias{sst_au_scan}
\title{scan}
\usage{
sst_au_scan(
  name,
  url,
  user_agent,
  http_auth_login = NULL,
  http_auth_pass = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{name}{(required) Name of your project}

\item{url}{(required) Page url to scan}

\item{user_agent}{(required)}

\item{http_auth_login}{(optional) Login for Basic HTTP authentication}

\item{http_auth_pass}{(optional) Password for Basic HTTP authentication}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
scan
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/3ca7uk9bdicd8-scan}{here}.
}

