% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_set_settings}
\alias{sst_au_set_settings}
\title{setSettings}
\usage{
sst_au_set_settings(
  project_id,
  main_settings,
  mail_trigger_settings,
  schedule_settings,
  scan_setting,
  dont_scan_keywords_block = NULL,
  only_scan_keywords_block = NULL,
  base_auth_block = NULL,
  errors_settings = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{main_settings}{(required)}

\item{mail_trigger_settings}{(required) Trigger mail settings}

\item{schedule_settings}{(required) Scan schedule settings}

\item{scan_setting}{(required) Scan type settings}

\item{dont_scan_keywords_block}{(optional) The Serpstat robot will exclude pages from crawling if the URL contains any of the specified words. This setting applies when the scan type is set to Sitemap too. Optional block, defaults to `\{checked: false, keywords: ""\}`.}

\item{only_scan_keywords_block}{(optional) The Serpstat robot will crawl pages with the following words in the URL. The setting is relevant for the scan type Sitemap too. Optional block, defaults to `\{checked: false, keywords: ""\}`.}

\item{base_auth_block}{(optional) If the site is restricted access, provide access for verification. Optional block, defaults to `\{login: "", password: ""\}`.}

\item{errors_settings}{(optional) Optional settings for error thresholds. If omitted, default values are used.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
setSettings
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/od01obqk4amtd-set-settings}{here}.
}

