% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_stop_page}
\alias{sst_au_stop_page}
\title{stop}
\usage{
sst_au_stop_page(
  page_id,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{page_id}{(required)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
stop
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/fghia9xpgzjzd-stop}{here}.
}

