% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{sst_au_user_log}
\alias{sst_au_user_log}
\title{userLog}
\usage{
sst_au_user_log(
  report_id = NULL,
  page_size = 100,
  page = 1,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{report_id}{(optional)}

\item{page_size}{(optional) Number of results per page in response}

\item{page}{(optional) page number for pagination}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
userLog
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/igusdr32uo9wi-user-log}{here}.
}

