% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_anchors}
\alias{sst_bl_anchors}
\title{Referring anchors}
\usage{
sst_bl_anchors(
  domain,
  search_type = "domain",
  anchor = NULL,
  count = NULL,
  sort = NULL,
  order = NULL,
  page = 1,
  size = 100,
  filter = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) A domain name to analyze.}

\item{search_type}{(optional) Default value is 'domain' for domain data only.}

\item{anchor}{(optional) Anchor text to filter by.}

\item{count}{(optional) Number of words in anchor to filter by.}

\item{sort}{(optional) A field to sort the response.}

\item{order}{(optional) The order of sorting.}

\item{page}{(optional) Response page number. Default is 1.}

\item{size}{(optional) Response page size. Default is 100.}

\item{filter}{(optional) The nested list of filtering options.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns anchors data.
}
\description{
Returns anchors of backlinks for the analyzed domain.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/3m88s123b99ic-get-anchors}{here}.
}

\section{API credits consumption}{
 1 per returned anchor.
}

