% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_changes_history}
\alias{sst_bl_changes_history}
\title{Backlinks changes history}
\usage{
sst_bl_changes_history(
  domain,
  search_type = "domain",
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) A domain name to analyze.}

\item{search_type}{(optional) Default value is 'domain' for domain data only.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns changes history data.
}
\description{
Returns history of backlinks changes (new and lost backlinks) over time.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/sfh2g1k2udjve-get-backlinks-changes-history}{here}.
}

\section{API credits consumption}{
 1 per request.
}

