% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_domains_intersection}
\alias{sst_bl_domains_intersection}
\title{Intersecting referring domains}
\usage{
sst_bl_domains_intersection(
  domain,
  compare_domains,
  sort = NULL,
  order = NULL,
  page = 1,
  size = 100,
  filter = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Target domain name.}

\item{compare_domains}{(required) A character vector of domains for comparison.}

\item{sort}{(optional) A field to sort the response.}

\item{order}{(optional) The order of sorting.}

\item{page}{(optional) Response page number. Default is 1.}

\item{size}{(optional) Response page size. Default is 100.}

\item{filter}{(optional) The nested list of filtering options.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns common referring domains data.
}
\description{
Returns referring domains that link to comparison domains but not to the target domain.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/3smzy5nr2892y-get-intersect}{here}.
}

\section{API credits consumption}{
 1 per each referring domain in response.
}

