% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_domains_intersection_summary}
\alias{sst_bl_domains_intersection_summary}
\title{Intersecting referring domains summary}
\usage{
sst_bl_domains_intersection_summary(
  domain,
  compare_domains,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Target domain name.}

\item{compare_domains}{(required) A character vector of domains to find common donors with.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns summary of common referring domains.
}
\description{
Returns summary of intersecting referring domains.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/3jfen5fhig0hr-get-intersect-summary}{here}.
}

\section{API credits consumption}{
 1 per request.
}

