% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_out_domains}
\alias{sst_bl_out_domains}
\title{Outgoing domains}
\usage{
sst_bl_out_domains(
  domain,
  search_type = "domain",
  sort = NULL,
  order = NULL,
  filter = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) A domain name to analyze.}

\item{search_type}{(optional) Default value is 'domain' for domain data only.}

\item{sort}{(optional) A field to sort the response.}

\item{order}{(optional) The order of sorting.}

\item{filter}{(optional) The nested list of filtering options.}

\item{page}{(optional) Response page number. Default is 1.}

\item{size}{(optional) Response page size. Default is 100.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns outgoing domains data.
}
\description{
Returns domains linked from the analyzed domain.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/xh16az2s7dl9u-get-out-domains}{here}.
}

\section{API credits consumption}{
 1 per returned outgoing domain.
}

