% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_redirected_domains}
\alias{sst_bl_redirected_domains}
\title{Redirected domains}
\usage{
sst_bl_redirected_domains(
  domain,
  sort = NULL,
  order = NULL,
  filter = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) A domain name to analyze.}

\item{sort}{(optional) A field to sort the response.}

\item{order}{(optional) The order of sorting.}

\item{filter}{(optional) The nested list of filtering options.}

\item{page}{(optional) Response page number. Default is 1.}

\item{size}{(optional) Response page size. Default is 100.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns redirected domains data.
}
\description{
Returns domains from which redirects were performed to the analyzed domain.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/9bxeyqwdvm3y7-get-redirected-domains}{here}.
}

\section{API credits consumption}{
 1 per returned redirected domain.
}

