% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_tld_distribution}
\alias{sst_bl_tld_distribution}
\title{TLD distribution}
\usage{
sst_bl_tld_distribution(
  domain,
  search_type = "domain",
  language = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) A domain name to analyze.}

\item{search_type}{(optional) Default value is 'domain' for domain data only.}

\item{language}{(optional) Language in which the country_name field is returned.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns TLD distribution data.
}
\description{
Returns distribution of top-level domains (TLDs) for the referring domains.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/texxxf7p0bq4a-get-distributiontld}{here}.
}

\section{API credits consumption}{
 1 per request.
}

