% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlinks.R
\name{sst_bl_top_anchors}
\alias{sst_bl_top_anchors}
\title{Top anchors}
\usage{
sst_bl_top_anchors(
  domain,
  search_type = "domain",
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) A domain name to analyze.}

\item{search_type}{(optional) Default value is 'domain' for domain data only.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns top anchors data.
}
\description{
Returns top anchors for the analyzed domain.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/r2y4zmc95iwyo-get-top-anchors}{here}.
}

\section{API credits consumption}{
 1 per request.
}

