% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credits.R
\name{sst_cr_get_audit_stats}
\alias{sst_cr_get_audit_stats}
\title{getAuditStats}
\usage{
sst_cr_get_audit_stats(
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getAuditStats
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/ulxuqx7ekbs4q-get-audit-stats}{here}.
}

