% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_classification.R
\name{sst_dc_add_task}
\alias{sst_dc_add_task}
\title{addTask}
\usage{
sst_dc_add_task(
  domains,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domains}{(required) An array of domain names}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
addTask
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/6bxxs5y9y3pf5-add-task}{here}.
}

