% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_classification.R
\name{sst_dc_get_task}
\alias{sst_dc_get_task}
\title{getTask}
\usage{
sst_dc_get_task(
  task_id,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{task_id}{(required) Task identifier}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getTask
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/cr2rhxq0g4qeh-get-task}{here}.
}

