% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_classification.R
\name{sst_dc_get_task_list}
\alias{sst_dc_get_task_list}
\title{getTaskList}
\usage{
sst_dc_get_task_list(
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{page}{(required) Page number in response}

\item{size}{(required) Number of results per page in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getTaskList
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/v4t7hbaw82qal-get-task-list}{here}.
}

