% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{sst_pm_create_project}
\alias{sst_pm_create_project}
\title{Create a new project}
\usage{
sst_pm_create_project(
  domain,
  name,
  groups = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Domain to get data for.}

\item{name}{(required) The name of the project. Can be different from the
domain.}

\item{groups}{(optional) A list of project groups the project should be added
to.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns the project ID for the created project.
}
\description{
Creates a new project in Serpstat.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/s04fhpr4grvek-create-project}{here}.
}

\section{API credits consumption}{
 1 project credit.
}

\examples{
\dontrun{
sst_pm_create_project(
  domain    = 'serpstat.com',
  name      = 'Serpstat'
  )$data$project_id
}
}
