% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{sst_pm_delete_project}
\alias{sst_pm_delete_project}
\title{Delete the existing project}
\usage{
sst_pm_delete_project(
  api_token = Sys.getenv("SERPSTAT_API_TOKEN"),
  project_id = NULL,
  return_method = "list"
)
}
\arguments{
\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}.Default is Sys.getenv('SERPSTAT_API_TOKEN').}

\item{project_id}{(required) ID of the project in Serpstat.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}
}
\value{
Returns the state of the deletion operation.
}
\description{
Deletes the existing project in Serpstat by project ID.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/z6pip7gt8wqo0-delete-project}{here}.
}

\section{API credits consumption}{
 returns 1 project credit.
}

\examples{
\dontrun{
sst_pm_delete_project(
  project_id = 12345
  )
}
}
