% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sst_return_check}
\alias{sst_return_check}
\title{Preprocess the API response}
\usage{
sst_return_check(response_content, return_method)
}
\arguments{
\item{response_content}{The result of \code{\link{sst_call_api_method}} call.}

\item{return_method}{Accepted values are 'list' to return data object as list
or 'df' to return data object as data.frame.}
}
\value{
response_content with a data object as list or data.frame.
}
\description{
Every API call returns a JSON object. This object is transformed to a list.
Depending on return_method parameter the data element of this list will be a
list or data.frame.
}
