% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_add_project_keywords}
\alias{sst_rt_add_project_keywords}
\title{addProjectKeywords}
\usage{
sst_rt_add_project_keywords(
  project_id,
  keywords,
  tags = NULL,
  url = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{keywords}{(required) Array of keywords}

\item{tags}{(optional) Tags for keywords (created automatically if don't exist)}

\item{url}{(optional) Target URL for rank checking}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
addProjectKeywords
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

