% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_add_rt_project}
\alias{sst_rt_add_rt_project}
\title{addRtProject}
\usage{
sst_rt_add_rt_project(
  name,
  domain,
  group = NULL,
  parsing_type = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{name}{(required) Project name}

\item{domain}{(required) Main domain to track}

\item{group}{(optional) Project group}

\item{parsing_type}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
addRtProject
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

