% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_create_and_run_project_tracking}
\alias{sst_rt_create_and_run_project_tracking}
\title{createAndRunProjectTracking}
\usage{
sst_rt_create_and_run_project_tracking(
  domain,
  keywords,
  name = NULL,
  group = NULL,
  regions = NULL,
  parsing_type = NULL,
  mirrors = NULL,
  schedule = NULL,
  competitors = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Main domain}

\item{keywords}{(required)}

\item{name}{(optional) Project name}

\item{group}{(optional) Project group}

\item{regions}{(optional) Tracking regions (auto-detected if not specified)}

\item{parsing_type}{(optional)}

\item{mirrors}{(optional) Domain mirrors (max. 10)}

\item{schedule}{(optional)}

\item{competitors}{(optional) Competitors (max. 10)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
createAndRunProjectTracking
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

