% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_delete_project_mirror}
\alias{sst_rt_delete_project_mirror}
\title{deleteProjectMirror}
\usage{
sst_rt_delete_project_mirror(
  project_id,
  mirror_uuid,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{mirror_uuid}{(required) Mirror UUID (from getProjectMirrors)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
deleteProjectMirror
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

