% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_get_project_keywords}
\alias{sst_rt_get_project_keywords}
\title{getProjectKeywords}
\usage{
sst_rt_get_project_keywords(
  project_id,
  keyword_ids = NULL,
  search_string = NULL,
  tags = NULL,
  page = 1,
  size = 100,
  sort = NULL,
  order = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{keyword_ids}{(optional) Filter by keyword ID array}

\item{search_string}{(optional)}

\item{tags}{(optional) Filter by tags (empty array = all tags)}

\item{page}{(optional)}

\item{size}{(optional)}

\item{sort}{(optional)}

\item{order}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getProjectKeywords
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

