% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_get_project_positions}
\alias{sst_rt_get_project_positions}
\title{getProjectPositions}
\usage{
sst_rt_get_project_positions(
  project_id,
  project_region_ids,
  domains = NULL,
  filter = NULL,
  page = 1,
  page_size = 100,
  date_from = NULL,
  date_to = NULL,
  sort = NULL,
  order = NULL,
  sort_region = NULL,
  tag_ids = NULL,
  keyword_ids = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{project_region_ids}{(required) Filter by project regions (get via getProjectRegions)}

\item{domains}{(optional) Filter by domains (UUID) - main + mirrors + competitors}

\item{filter}{(optional)}

\item{page}{(optional)}

\item{page_size}{(optional)}

\item{date_from}{(optional)}

\item{date_to}{(optional)}

\item{sort}{(optional)}

\item{order}{(optional)}

\item{sort_region}{(optional)}

\item{tag_ids}{(optional)}

\item{keyword_ids}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getProjectPositions
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

