% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_run_project_tracking}
\alias{sst_rt_run_project_tracking}
\title{runProjectTracking}
\usage{
sst_rt_run_project_tracking(
  project_id,
  keyword_ids = NULL,
  region_ids = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{keyword_ids}{(optional) Keyword IDs (if empty, runs for all keywords)}

\item{region_ids}{(optional) Region IDs (if empty, runs for all regions)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
runProjectTracking
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

