% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tracker.R
\name{sst_rt_update_project_keywords}
\alias{sst_rt_update_project_keywords}
\title{updateProjectKeywords}
\usage{
sst_rt_update_project_keywords(
  project_id,
  keywords,
  regions = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{project_id}{(required)}

\item{keywords}{(required) List of keywords to update. From 1 to 100 elements.}

\item{regions}{(optional) List of project region IDs to apply the update to. If not provided or all provided IDs do not belong to the project, the update applies to all project regions. Max 100 elements.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
updateProjectKeywords
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jenasqbwtxdlr-introduction-to-serpstat-api}{here}.
}

