% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domain_ad_competitors}
\alias{sst_sa_domain_ad_competitors}
\title{Domain advertising competitors}
\usage{
sst_sa_domain_ad_competitors(
  domain,
  se,
  sort = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Domain name to analyze.}

\item{se}{(required) Search engine alias (db_name) returned by
\code{\link{sst_sa_database_info}}.}

\item{sort}{(optional) A field to sort the response. See Sorting for more
details.}

\item{page}{(optional) Response page number if there are many pages in response.}

\item{size}{(optional) Response page size.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns a list of PPC competitors with their metrics.
}
\description{
Returns competing domains in paid search (PPC) results.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/g2arx1ss4o34o-get-ads-competitors}{here}.
}

\section{API credits consumption}{
 1 per returned competitor.
}

\section{Sorting}{

  You can sort the response using \code{sort} argument. It must be a list
  with a single named element. The name of the element must match one of
  parameters in response. The value of the element must be \code{asc} for
  ascending order and \code{desc} for descending order. For example,
  \code{sort = list(ads = 'desc')} would sort the response by \code{ads}
  parameter in descending order.
}

\examples{
\dontrun{
sst_sa_domain_ad_competitors(
  domain        = 'serpstat.com',
  se            = 'g_us',
  page          = 1,
  size          = 100,
  return_method = 'df'
)$data
}
}
