% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domain_aio_brand_opportunities}
\alias{sst_sa_domain_aio_brand_opportunities}
\title{AI Overview Brand Opportunities}
\usage{
sst_sa_domain_aio_brand_opportunities(
  domains,
  brand,
  search_engine = "g_us",
  page = 1,
  page_size = 100,
  sort = NULL,
  order = NULL,
  filters = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domains}{(required) Vector of domain names (up to 100) to analyze.}

\item{brand}{(required) Brand name to search for in AIO texts.}

\item{search_engine}{(optional) Search engine database. Allowed values: g_us, g_ua. Default: g_us.}

\item{page}{(optional) Response page number.}

\item{page_size}{(optional) Response page size. Default: 100.}

\item{sort}{(optional) Sort field. Allowed values: keyword. Default is NULL.}

\item{order}{(optional) Sort direction. Allowed values: asc, desc. Default is NULL.}

\item{filters}{(optional) A list of filtering options. See Filtering for more
details.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns a paginated list of keywords with AIO data.
}
\description{
Returns a paginated list of keywords with AI Overview (AIO) data for a given brand and list of domains.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/aio_brand_opportunities_method-get-aio-brand-opportunities}{here}.
}

\section{API credits consumption}{
 2 credits per row in response.
}

\section{Filtering}{

  To filter the results you can use \code{filters} argument. It must be a
  list of named elements. The name of the element must match one of the
  filtering parameters. See API docs for more details. For example,
  \code{filters = list(queries_from = 0, queries_to = 10)} would narrow
  the results to include only the keywords that have a search volume
  between 0 and 10.
}

\examples{
\dontrun{
sst_sa_domain_aio_brand_opportunities(
  domains       = c('amazon.com'),
  brand         = 'Amazon',
  search_engine = 'g_us',
  page          = 1,
  page_size     = 100,
  return_method = 'df'
)$data
}
}
