% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domain_aio_summary}
\alias{sst_sa_domain_aio_summary}
\title{AI Overview Brand Summary}
\usage{
sst_sa_domain_aio_summary(
  domains,
  brand,
  search_engine = "g_us",
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domains}{(required) Vector of domain names (up to 100) to analyze.}

\item{brand}{(required) Brand name to search for in AIO texts.}

\item{search_engine}{(optional) Search engine database. Allowed values: g_us, g_ua. Default: g_us.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns a summary of AIO brand presence.
}
\description{
Returns a summary of AI Overview (AIO) presence for a given brand and list of domains.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/aio_summary_method-get-aio-summary}{here}.
}

\section{API credits consumption}{
 5 credits per request.
}

\examples{
\dontrun{
sst_sa_domain_aio_summary(
  domains       = c('amazon.com'),
  brand         = 'Amazon',
  search_engine = 'g_us',
  return_method = 'df'
)$data
}
}
