% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domain_all_regions_traffic}
\alias{sst_sa_domain_all_regions_traffic}
\title{Traffic in all regions for a domain}
\usage{
sst_sa_domain_all_regions_traffic(
  domain,
  sort = c("traff"),
  order = "desc",
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Domain name to analyze.}

\item{sort}{(optional) Vector of sorting fields. Default is "traff".}

\item{order}{(optional) Order of sorting. Allowed values: asc, desc. Default: desc.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns traffic data for all regions.
}
\description{
Returns search traffic info for a domain across all databases. Databases with no keywords are not shown.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/qa2ih3uak03yo-get-all-regions-traffic}{here}.
}

\section{API credits consumption}{
 1 per region in response.
}

\examples{
\dontrun{
sst_sa_domain_all_regions_traffic(
  domain        = 'serpstat.com',
  sort          = c('traff'),
  order         = 'desc',
  return_method = 'df'
)$data
}
}
