% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domain_keywords_by_language}
\alias{sst_sa_domain_keywords_by_language}
\title{Domain keywords by language}
\usage{
sst_sa_domain_keywords_by_language(
  domain,
  se,
  with_subdomains = TRUE,
  url = NULL,
  keywords = NULL,
  minus_keywords = NULL,
  sort = NULL,
  filters = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domain}{(required) Domain name to get data for.}

\item{se}{(required) Search engine alias (db_name) returned by
\code{\link{sst_sa_database_info}}.}

\item{with_subdomains}{(optional) Retrieve data for subdomains too. Default is TRUE.}

\item{url}{(optional) Get the results for this URL only.}

\item{keywords}{(optional) A vector of words. Keywords in response will
contain these words.}

\item{minus_keywords}{(optional) A vector of words. Keywords in response will
not contain these words.}

\item{sort}{(optional) A field to sort the response. See Sorting for more
details.}

\item{filters}{(optional) A list of filtering options. See Filtering for more
details.}

\item{page}{(optional) Response page number if there are many pages in response.}

\item{size}{(optional) Response page size.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns keywords with language detection info.
}
\description{
Returns keywords which the analyzed domain ranks for, enriched with the detected language of each keyword.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/ke5u46gfs2a0d-get-domain-keywords-by-language}{here}.
}

\section{API credits consumption}{
 1 credit per member of data object array in response.
}

\section{Sorting}{

  You can sort the response using \code{sort} argument. It must be a list
  with a single named element. The name of the element must match one of
  parameters in response. The value of the element must be \code{asc} for
  ascending order and \code{desc} for descending order. For example,
  \code{sort = list(ads = 'desc')} would sort the response by \code{ads}
  parameter in descending order.
}

\section{Filtering}{

  To filter the results you can use \code{filters} argument. It must be a
  list of named elements. The name of the element must match one of the
  filtering parameters. See API docs for more details. For example,
  \code{filters = list(queries_from = 0, queries_to = 10)} would narrow
  the results to include only the keywords that have a search volume
  between 0 and 10.
}

\examples{
\dontrun{
sst_sa_domain_keywords_by_language(
  domain        = 'serpstat.com',
  se            = 'g_us',
  page          = 1,
  size          = 100,
  return_method = 'df'
)$data
}
}
