% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domains_rating_data}
\alias{sst_sa_domains_rating_data}
\title{Rating Data for Domains}
\usage{
sst_sa_domains_rating_data(
  domains,
  se,
  sort = NULL,
  order = NULL,
  filters = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domains}{(required) Vector of domain names (up to 200) to analyze.}

\item{se}{(required) Search engine alias (db_name) returned by
\code{\link{sst_sa_database_info}}.}

\item{sort}{(optional) Sort field name.}

\item{order}{(optional) Sort order.}

\item{filters}{(optional) A list of filtering options. See Filtering for more
details.}

\item{page}{(optional) Response page number if there are many pages in response.}

\item{size}{(optional) Response page size.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns market research ratings for the domains.
}
\description{
Returns market research data for a list of specified domains, including their category classification and SEO metrics.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/domains_rating_data_method-get-domains-rating-data}{here}.
}

\section{API credits consumption}{
 1 credit per row in response.
}

\section{Filtering}{

  To filter the results you can use \code{filters} argument. It must be a
  list of named elements. The name of the element must match one of the
  filtering parameters. See API docs for more details. For example,
  \code{filters = list(queries_from = 0, queries_to = 10)} would narrow
  the results to include only the keywords that have a search volume
  between 0 and 10.
}

\examples{
\dontrun{
sst_sa_domains_rating_data(
  domains       = c('amazon.com', 'ebay.com'),
  se            = 'g_us',
  page          = 1,
  size          = 100,
  return_method = 'df'
)$data
}
}
