% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_domains_unique_keywords}
\alias{sst_sa_domains_unique_keywords}
\title{Unique keywords of domains}
\usage{
sst_sa_domains_unique_keywords(
  domains,
  minus_domain,
  se,
  filters = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{domains}{(required) Vector of domain names (min 1, max 2) to analyze.}

\item{minus_domain}{(required) Domain name with keywords to exclude.}

\item{se}{(required) Search engine alias (db_name) returned by
\code{\link{sst_sa_database_info}}.}

\item{filters}{(optional) A list of filtering options. See Filtering for more
details.}

\item{page}{(optional) Response page number if there are many pages in response.}

\item{size}{(optional) Response page size.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns unique keywords for the domains.
}
\description{
Returns unique keywords of one or two domains that a third domain does not rank for.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/zerum86tpb5xe-get-domains-uniq-keywords}{here}.
}

\section{API credits consumption}{
 1 per returned keyword.
}

\section{Filtering}{

  To filter the results you can use \code{filters} argument. It must be a
  list of named elements. The name of the element must match one of the
  filtering parameters. See API docs for more details. For example,
  \code{filters = list(queries_from = 0, queries_to = 10)} would narrow
  the results to include only the keywords that have a search volume
  between 0 and 10.
}

\examples{
\dontrun{
sst_sa_domains_unique_keywords(
  domains       = c('amazon.com'),
  minus_domain  = 'ebay.com',
  se            = 'g_us',
  page          = 1,
  size          = 100,
  return_method = 'df'
)$data
}
}
