% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_keyword_export_keywords_phrase}
\alias{sst_sa_keyword_export_keywords_phrase}
\title{exportKeywordsPhrase}
\usage{
sst_sa_keyword_export_keywords_phrase(
  keyword,
  search_engine,
  filters = NULL,
  sort = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keyword}{(required) Keyword to search for}

\item{search_engine}{(required)}

\item{filters}{(optional) Filters for search. Fields are combined using the **AND** logic. Numeric range fields (`_from` and `_to`) allow specifying minimum and maximum values. List fields (`_contain` and `_not_contain`) specify inclusion or exclusion criteria}

\item{sort}{(optional) Order of sorting the results in the format: `field: order`

- **field** — field to sort by (any field in response)
- **order** — sort direction (`asc` — ascending, `desc` — descending)

Example `\{"region_queries_count": "desc"\}`}

\item{page}{(optional) Page number in response}

\item{size}{(optional) Page number in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
exportKeywordsPhrase
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/bo7054tzxr861-export-keywords-phrase}{here}.
}

