% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_keyword_get_ad_keywords}
\alias{sst_sa_keyword_get_ad_keywords}
\title{getAdKeywords}
\usage{
sst_sa_keyword_get_ad_keywords(
  keyword,
  search_engine,
  domains,
  minus_keywords = NULL,
  filter = NULL,
  sort = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keyword}{(required) Keyword to search for}

\item{search_engine}{(required)}

\item{domains}{(required) Domain names list

Example: `["apple.com","verizon.com"]`}

\item{minus_keywords}{(optional) List of keywords to exclude from the search}

\item{filter}{(optional) Filters for search. Fields are combined using the **AND** logic. Numeric range fields (`_from` and `_to`) allow specifying minimum and maximum values. List fields (`_contain` and `_not_contain`) specify inclusion or exclusion criteria}

\item{sort}{(optional) Order of sorting the results in the format: `field: order`

- **field** — field to sort by (any field in *data* section of response)
- **order** — sort direction (`asc` — ascending, `desc` — descending)

Example `\{"region_queries_count": "desc"\}`}

\item{page}{(optional) Page number in response}

\item{size}{(optional) Number of results per page in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getAdKeywords
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/iwm7yvtdj0d1u-get-ad-keywords}{here}.
}

