% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_keyword_get_ads_competitors}
\alias{sst_sa_keyword_get_ads_competitors}
\title{getAdsCompetitors}
\usage{
sst_sa_keyword_get_ads_competitors(
  keyword,
  search_engine,
  sort = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keyword}{(required) Keyword to search for}

\item{search_engine}{(required)}

\item{sort}{(optional) Order of sorting the results in the format: `field: order`

- **field** — field to sort by (any field in *domain detailed information* section of response)
- **order** — sort direction (`asc` — ascending, `desc` — descending)

Example `\{"region_queries_count": "desc"\}`}

\item{page}{(optional) Page number in response}

\item{size}{(optional) Number of results per page in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getAdsCompetitors
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/thvz04d3alhkz-get-ads-competitors}{here}.
}

