% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_keyword_get_competitors}
\alias{sst_sa_keyword_get_competitors}
\title{getCompetitors}
\usage{
sst_sa_keyword_get_competitors(
  keyword,
  search_engine,
  filters = NULL,
  sort = NULL,
  order = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keyword}{(required) Keyword to search for}

\item{search_engine}{(required) Search engine short name. Example: `"search_engine": "g_us"`. Refer to full list of [search engines](../serpstat-public-api/ba97ni814ao9p-search-engine-short-names)}

\item{filters}{(optional) Filters for search. Fields are combined using the **AND** logic. Numeric range fields (`_from` and `_to`) allow specifying minimum and maximum values.}

\item{sort}{(optional) Field to sort by}

\item{order}{(optional) Sort direction}

\item{page}{(optional) Page number}

\item{size}{(optional) Number of results per page}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getCompetitors
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/kw9p2xm4cv3nf-get-competitorsv2}{here}.
}

