% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_keyword_get_top_urls}
\alias{sst_sa_keyword_get_top_urls}
\title{getTopUrls}
\usage{
sst_sa_keyword_get_top_urls(
  keyword,
  search_engine,
  sort = NULL,
  order = NULL,
  page = 1,
  page_size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keyword}{(required) Keyword to search for}

\item{search_engine}{(required)}

\item{sort}{(optional) Sorting by parameters(Any field in *urls* section of response)}

\item{order}{(optional) Sorting order}

\item{page}{(optional) Page number}

\item{page_size}{(optional) Number of results per page}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getTopUrls
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/hwiqar7egodd6-get-top-urls}{here}.
}

