% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_keyword_trends}
\alias{sst_sa_keyword_trends}
\title{Keyword trends}
\usage{
sst_sa_keyword_trends(
  keywords,
  se,
  months = 48,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keywords}{(required) A vector of keywords to analyze.}

\item{se}{(required) Search engine alias (db_name) returned by
\code{\link{sst_sa_database_info}}.}

\item{months}{(optional) Number of months to get data for (up to 48). Default is 48.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns monthly search volume trends for each keyword.
}
\description{
Returns monthly search volume trends for a list of keywords over a specified period (up to 48 months).
You can send up to 100 keywords per request.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/kw3t8rn5vx1qm-get-keyword-trends}{here}.
}

\section{API credits consumption}{
 10 credits per keyword that returned data. If no keywords returned data — 1 credit is charged for the whole request.
}

\examples{
\dontrun{
sst_sa_keyword_trends(
  keywords      = c('seo', 'ppc', 'serpstat'),
  se            = 'g_us',
  months        = 12,
  return_method = 'df'
)$data
}
}
