% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_market_categories}
\alias{sst_sa_market_categories}
\title{Market Research Categories}
\usage{
sst_sa_market_categories(
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token from
\href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').}
}
\value{
Returns a list of category identifiers and names.
}
\description{
Returns a complete list of available market research categories used for domain classification.
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/market_categories_method-get-market-categories}{here}.
}

\section{API credits consumption}{
 0 credits.
}

\examples{
\dontrun{
sst_sa_market_categories(return_method = 'df')$data
}
}
