% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_url_get_summary_traffic}
\alias{sst_sa_url_get_summary_traffic}
\title{getSummaryTraffic}
\usage{
sst_sa_url_get_summary_traffic(
  search_engine,
  domain,
  url_contains,
  output_data = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{search_engine}{(required)}

\item{domain}{(required) The domain for which to retrieve traffic and keyword data}

\item{url_contains}{(required) Searched part of URL}

\item{output_data}{(optional) Output data value.  Both options enable by default, choose one to spend less credits.}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getSummaryTraffic
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/ukf1sgnx5mop4-get-summary-traffic}{here}.
}

