% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_url_get_url_competitors}
\alias{sst_sa_url_get_url_competitors}
\title{getUrlCompetitors}
\usage{
sst_sa_url_get_url_competitors(
  search_engine,
  url,
  sort = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{search_engine}{(required)}

\item{url}{(required) URL for finding competing URLs}

\item{sort}{(optional) Order of sorting the results}

\item{page}{(optional) Page number in response}

\item{size}{(optional) Number of results per page in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getUrlCompetitors
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/0euusensv6kpf-get-url-competitors}{here}.
}

