% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_url_get_url_keywords}
\alias{sst_sa_url_get_url_keywords}
\title{getUrlKeywords}
\usage{
sst_sa_url_get_url_keywords(
  search_engine,
  url,
  with_intents = NULL,
  sort = NULL,
  filters = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{search_engine}{(required)}

\item{url}{(required) Analyzed page URL}

\item{with_intents}{(optional) Keyword intent

*This parameter works for `g_ua` and `g_us` database only*}

\item{sort}{(optional) Order of sorting the results}

\item{filters}{(optional) Filters for search. Fields are combined using the "AND" logic. Numeric range fields (`_from` and `_to`) allow specifying minimum and maximum values. List fields (`_contain` and `_not_contain`) specify inclusion or exclusion criteria}

\item{page}{(optional) Page number in response}

\item{size}{(optional) Number of results per page in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getUrlKeywords
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/58bzxr2rb7qwc-get-url-keywords}{here}.
}

