% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_analytics.R
\name{sst_sa_url_get_url_missing_keywords}
\alias{sst_sa_url_get_url_missing_keywords}
\title{getUrlMissingKeywords}
\usage{
sst_sa_url_get_url_missing_keywords(
  url,
  search_engine,
  sort = NULL,
  filters = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{url}{(required) Analyzed URL}

\item{search_engine}{(required)}

\item{sort}{(optional) Order of sorting the results}

\item{filters}{(optional) Filters for search. Fields are combined using the "AND" logic. Numeric range fields (`_from` and `_to`) allow specifying minimum and maximum values. List fields (`_contain` and `_not_contain`) specify inclusion or exclusion criteria}

\item{page}{(optional) Page number in response}

\item{size}{(optional) Number of results per page in response}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getUrlMissingKeywords
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/6d8wt6v89v2ap-get-url-missing-keywords}{here}.
}

