% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_crawling.R
\name{sst_sc_add_keyword_list}
\alias{sst_sc_add_keyword_list}
\title{addKeywordList}
\usage{
sst_sc_add_keyword_list(
  keywords,
  search_engine_id,
  country_id,
  region_id = NULL,
  language_id = NULL,
  type_id = NULL,
  type = NULL,
  pages = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keywords}{(required) Keywords with commas for parsing.

Example: `["samsung, iphone", "nike, adidas"]`}

\item{search_engine_id}{(required) Search engine identifier

* **1** — Google}

\item{country_id}{(required)}

\item{region_id}{(optional)}

\item{language_id}{(optional)}

\item{type_id}{(optional)}

\item{type}{(optional)}

\item{pages}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
addKeywordList
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/9elengxj24hwb-add-keyword-list}{here}.
}

