% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_crawling.R
\name{sst_sc_add_task}
\alias{sst_sc_add_task}
\title{Add crawling task}
\usage{
sst_sc_add_task(
  keywords,
  country_id,
  se_id = 1,
  region_id = NULL,
  language_id = 1,
  device_type_id = 1,
  results_type = "regular_aio",
  pages = 1,
  api_token = Sys.getenv("SERPSTAT_API_TOKEN"),
  return_method = "list"
)
}
\arguments{
\item{keywords}{(required) A character vector or comma-separated string of
keywords.}

\item{country_id}{(required) Country identifier (e.g. 23 for United States).
Range: 1-247.}

\item{se_id}{(optional) Search engine ID. Default is 1 (Google).}

\item{region_id}{(optional) Region or city identifier.}

\item{language_id}{(optional) Language ID. Default is 1 (English).
Range: 1-48.}

\item{device_type_id}{(optional) Device type ID. Default is 1 (Desktop).
Range: 1-2.}

\item{results_type}{(optional) Type of Google results. Default is
"regular_aio". Allowed values: "regular", "local", "regular_aio".}

\item{pages}{(optional) Number of SERP pages to parse. Default is 1.
Range: 1-10.}

\item{api_token}{(optional) Serpstat API token. Default is
Sys.getenv("SERPSTAT_API_TOKEN").}

\item{return_method}{(optional) Accepted values are "list" (default) to
return data object as list or "df" to return data object as data.frame.}
}
\value{
Returns the task ID.
}
\description{
Submits keywords to crawl regular SERP and local results.
}
\section{API docs}{

 Check all the values for request and response fields
 \href{https://api-docs.serpstat.com/docs/serpstat-public-api/fd78zl20ap9w1-add-task}{here}.
}

\section{API credits consumption}{

 Depends on the number of keywords and pages to parse.
}

\examples{
\dontrun{
sst_sc_add_task(
  keywords = c("explain the gauss rifle principle",
               "why iphone better than samsung"),
  country_id = 23
)
}
}
