% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_crawling.R
\name{sst_sc_get_keyword_serp}
\alias{sst_sc_get_keyword_serp}
\title{getKeywordSerp}
\usage{
sst_sc_get_keyword_serp(
  task_id,
  keyword_id,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{task_id}{(required)}

\item{keyword_id}{(required) Id number of a required keyword for getting a raw SERP (you can get it in the SERP crawling response from the method `tasks.getTaskResult`)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getKeywordSerp
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/bfwx6sfgzlioz-get-keyword-serp}{here}.
}

