% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_crawling.R
\name{sst_sc_get_list}
\alias{sst_sc_get_list}
\title{getList}
\usage{
sst_sc_get_list(
  page = 1,
  page_size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{page}{(optional) Page Number}

\item{page_size}{(optional) Page size}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getList
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/2jkkku527phy7-get-list}{here}.
}

