% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_crawling.R
\name{sst_sc_get_parsing_balance}
\alias{sst_sc_get_parsing_balance}
\title{getParsingBalance}
\usage{
sst_sc_get_parsing_balance(
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getParsingBalance
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/hx1b8q97pwu9y-get-parsing-balance}{here}.
}

