% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_crawling.R
\name{sst_sc_get_task_result}
\alias{sst_sc_get_task_result}
\title{Get crawling task result}
\usage{
sst_sc_get_task_result(
  task_id,
  page = 1,
  api_token = Sys.getenv("SERPSTAT_API_TOKEN"),
  return_method = "list"
)
}
\arguments{
\item{task_id}{(required) Crawling identifier.}

\item{page}{(optional) Page number for pagination. Results are grouped in
pages of 100 keywords. Default is 1.}

\item{api_token}{(optional) Serpstat API token. Default is
Sys.getenv("SERPSTAT_API_TOKEN").}

\item{return_method}{(optional) Accepted values are "list" (default) to
return data object as list or "df" to return data object as data.frame.}
}
\value{
Returns the parsing results for the specified task.
}
\description{
Retrieves parsing results of a crawling task.
}
\section{API docs}{

 Check all the values for request and response fields
 \href{https://api-docs.serpstat.com/docs/serpstat-public-api/50bwxkycuia8p-get-task-result}{here}.
}

\section{API credits consumption}{
 0
}

\examples{
\dontrun{
sst_sc_get_task_result(
  task_id = 5484945
)
}
}
