% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_management.R
\name{sst_tm_add_user}
\alias{sst_tm_add_user}
\title{addUser}
\usage{
sst_tm_add_user(
  email,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{email}{(required) user email}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
addUser
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/eugask0z2vg4e-add-user}{here}.
}

