% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_management.R
\name{sst_tm_deactivate_user}
\alias{sst_tm_deactivate_user}
\title{deactivateUser}
\usage{
sst_tm_deactivate_user(
  user_id,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{user_id}{(required) User identifier at Serpstat. Can be obtained via `TeamManagement.getList` method}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
deactivateUser
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/b0ui1egjoh609-deactivate-user}{here}.
}

