% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_management.R
\name{sst_tm_get_list}
\alias{sst_tm_get_list}
\title{getList}
\usage{
sst_tm_get_list(
  search = NULL,
  page = 1,
  size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{search}{(optional) Optional part of email of member you need to find}

\item{page}{(optional) page number for pagination}

\item{size}{(optional) records per page}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getList
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/9565ktu8y7f0f-get-list}{here}.
}

