% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_management.R
\name{sst_tm_remove_user}
\alias{sst_tm_remove_user}
\title{removeUser}
\usage{
sst_tm_remove_user(
  user_id,
  merge_projects,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{user_id}{(required) User id at Serpstat. Can be found using `TeamManagement.getList` method.}

\item{merge_projects}{(required) `true` - transfer user's projects to the team owner.
`false` - projects will be deleted}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
removeUser
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/1jznkuymvwuh7-remove-user}{here}.
}

