% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_checker.R
\name{sst_vc_add_keyword_list_freq}
\alias{sst_vc_add_keyword_list_freq}
\title{addKeywordListFreq}
\usage{
sst_vc_add_keyword_list_freq(
  keywords,
  region_id,
  type_id = NULL,
  search_engine_id = NULL,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{keywords}{(required) Keywords to get metrics (max 50000)}

\item{region_id}{(required)}

\item{type_id}{(optional)}

\item{search_engine_id}{(optional)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
addKeywordListFreq
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/c18po8yuu2j06-add-keyword-list-freq}{here}.
}

