% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_checker.R
\name{sst_vc_get_task_result}
\alias{sst_vc_get_task_result}
\title{getTaskResult}
\usage{
sst_vc_get_task_result(
  task_id,
  page = 1,
  page_size = 100,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{task_id}{(required)}

\item{page}{(optional) Page number}

\item{page_size}{(optional) Number of results per a page}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getTaskResult
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/v91plc4ro30g6-get-task-result}{here}.
}

