% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_checker.R
\name{sst_vc_get_task_status}
\alias{sst_vc_get_task_status}
\title{getTaskStatus}
\usage{
sst_vc_get_task_status(
  task_id,
  return_method = "list",
  api_token = Sys.getenv("SERPSTAT_API_TOKEN")
)
}
\arguments{
\item{task_id}{(required)}

\item{return_method}{(optional) Accepted values are 'list' (default) to
return data object as list or 'df' to return data object as data.frame.}

\item{api_token}{(required) Serpstat API token.}
}
\value{
Returns the API response.
}
\description{
getTaskStatus
}
\section{API docs}{

 Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/ysi42jt8k737w-get-task-status}{here}.
}

